+++ core/src/main/java/org/eqasim/core/scenario/cutter/RunScenarioCutter.java
@@ -2,8 +2,9 @@ package org.eqasim.core.scenario.cutter;
 
 import java.io.File;
 import java.io.IOException;
-import java.net.MalformedURLException;
+import java.util.Collection;
 import java.util.Optional;
+import java.util.Set;
 
 import org.eqasim.core.components.travel_time.RecordedTravelTime;
 import org.eqasim.core.misc.InjectorBuilder;
@@ -16,6 +17,7 @@ import org.eqasim.core.scenario.cutter.network.NetworkCutter;
 import org.eqasim.core.scenario.cutter.network.RoadNetwork;
 import org.eqasim.core.scenario.cutter.outside.OutsideActivityAdapter;
 import org.eqasim.core.scenario.cutter.population.CleanHouseholds;
+import org.eqasim.core.scenario.cutter.population.CleanVehicles;
 import org.eqasim.core.scenario.cutter.population.PopulationCutter;
 import org.eqasim.core.scenario.cutter.population.PopulationCutterModule;
 import org.eqasim.core.scenario.cutter.population.RemoveEmptyPlans;
@@ -26,10 +28,12 @@ import org.eqasim.core.scenario.cutter.transit.TransitVehiclesCutter;
 import org.eqasim.core.scenario.routing.PopulationRouter;
 import org.eqasim.core.scenario.routing.PopulationRouterModule;
 import org.eqasim.core.scenario.validation.ScenarioValidator;
+import org.eqasim.core.scenario.validation.VehiclesValidator;
 import org.eqasim.core.simulation.EqasimConfigurator;
 import org.eqasim.core.simulation.mode_choice.AbstractEqasimExtension;
 import org.eqasim.core.simulation.termination.EqasimTerminationModule;
 import org.matsim.api.core.v01.Scenario;
+import org.matsim.contribs.discrete_mode_choice.modules.DiscreteModeChoiceModule;
 import org.matsim.core.config.CommandLine;
 import org.matsim.core.config.CommandLine.ConfigurationException;
 import org.matsim.core.config.Config;
@@ -40,11 +44,15 @@ import org.matsim.core.utils.timing.TimeInterpretationModule;
 import com.google.inject.Injector;
 
 public class RunScenarioCutter {
-	static public void main(String[] args)
-			throws ConfigurationException, MalformedURLException, IOException, InterruptedException {
+
+	public static final Collection<String> REQUIRED_ARGS = Set.of("config-path", "output-path", "extent-path");
+	public static final Collection<String> OPTIONAL_ARGS = Set.of("threads", "prefix", "extent-attribute",
+			"extent-value", "plans-path", "events-path", "skip-routing");
+
+	static public void main(String[] args) throws ConfigurationException, IOException, InterruptedException {
 		CommandLine cmd = new CommandLine.Builder(args) //
-				.requireOptions("config-path", "output-path", "extent-path") //
-				.allowOptions("threads", "prefix", "extent-attribute", "extent-value", "plans-path", "events-path") //
+				.requireOptions(REQUIRED_ARGS) //
+				.allowOptions(OPTIONAL_ARGS) //
 				.build();
 
 		// Load some configuration
@@ -64,9 +72,11 @@ public class RunScenarioCutter {
 		// Load scenario
 		EqasimConfigurator configurator = new EqasimConfigurator();
 		configurator.getModules().removeIf(m -> m instanceof EqasimTerminationModule);
-		
+
 		Config config = ConfigUtils.loadConfig(cmd.getOptionStrict("config-path"), configurator.getConfigGroups());
 		cmd.applyConfiguration(config);
+
+		VehiclesValidator.validate(config);
 
 		Optional<String> plansPath = cmd.getOption("plans-path");
 
@@ -101,7 +111,8 @@ public class RunScenarioCutter {
 
 		// Cut population
 		Injector populationCutterInjector = new InjectorBuilder(scenario) //
-				.addOverridingModules(configurator.getModules().stream().filter(module -> !(module instanceof AbstractEqasimExtension)).toList()) //
+				.addOverridingModules(configurator.getModules().stream()
+						.filter(module -> !(module instanceof AbstractEqasimExtension) && !(module instanceof DiscreteModeChoiceModule)).toList()) //
 				.addOverridingModule(
 						new PopulationCutterModule(extent, numberOfThreads, 40, cmd.getOption("events-path"))) //
 				.addOverridingModule(new CutterTravelTimeModule(travelTime)) //
@@ -122,6 +133,10 @@ public class RunScenarioCutter {
 		// ... and make households consistent
 		CleanHouseholds cleanHouseholds = new CleanHouseholds(scenario.getPopulation());
 		cleanHouseholds.run(scenario.getHouseholds());
+
+		// .. and make vehicles consistent
+		CleanVehicles cleanVehicles = new CleanVehicles(scenario.getPopulation());
+		cleanVehicles.run(scenario.getVehicles());
 
 		// Cut transit
 		StopSequenceCrossingPointFinder stopSequenceCrossingPointFinder = new DefaultStopSequenceCrossingPointFinder(
@@ -154,17 +169,21 @@ public class RunScenarioCutter {
 
 		// Final routing
 		Injector routingInjector = new InjectorBuilder(scenario) //
-				.addOverridingModules(configurator.getModules().stream().filter(module -> !(module instanceof AbstractEqasimExtension)).toList()) //
+				.addOverridingModules(configurator.getModules().stream()
+						.filter(module -> !(module instanceof AbstractEqasimExtension) && !(module instanceof DiscreteModeChoiceModule)).toList()) //
 				.addOverridingModule(new PopulationRouterModule(numberOfThreads, 100, false)) //
 				.addOverridingModule(new CutterTravelTimeModule(travelTime)) //
 				.addOverridingModule(new TimeInterpretationModule()) //
 				.build();
 
-		PopulationRouter router = routingInjector.getInstance(PopulationRouter.class);
-		router.run(scenario.getPopulation());
+		boolean skipRouting = Boolean.parseBoolean(cmd.getOption("skip-routing").orElse("false"));
 
-		// Check validity after cutting
-		scenarioValidator.checkScenario(scenario);
+		if (!skipRouting) {
+			PopulationRouter router = routingInjector.getInstance(PopulationRouter.class);
+			router.run(scenario.getPopulation());
+			// Check validity after cutting
+			scenarioValidator.checkScenario(scenario);
+		}
 
 		// Write scenario
 		ScenarioWriter scenarioWriter = new ScenarioWriter(config, scenario, prefix);