+++ core/src/main/java/org/eqasim/core/simulation/modes/feeder_drt/config/FeederDrtConfigGroup.java
@@ -2,15 +2,26 @@ package org.eqasim.core.simulation.modes.feeder_drt.config;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+import org.eqasim.core.simulation.modes.feeder_drt.router.access_egress_stop_search.CompositeAccessEgressStopSearchParameterSet;
+import org.eqasim.core.simulation.modes.feeder_drt.router.access_egress_stop_search.TransitStopByIdAccessEgressStopSearchParameterSet;
+import org.eqasim.core.simulation.modes.feeder_drt.router.access_egress_stop_search.TransitStopByModeAccessEgressStopSearchParameterSet;
+import org.matsim.contrib.common.util.ReflectiveConfigGroupWithConfigurableParameterSets;
 import org.matsim.contrib.dvrp.run.Modal;
-import org.matsim.contrib.util.ReflectiveConfigGroupWithConfigurableParameterSets;
 
 public class FeederDrtConfigGroup extends ReflectiveConfigGroupWithConfigurableParameterSets implements Modal {
     public static final String GROUP_NAME = "feederDrt";
     public FeederDrtConfigGroup() {
         super(GROUP_NAME);
-        this.addDefinition(AccessEgressStopSelectorParams.NAME, AccessEgressStopSelectorParams::new, () -> this.accessEgressStopSelectorConfig,
-                params -> this.accessEgressStopSelectorConfig = (AccessEgressStopSelectorParams) params);
+
+        //AccessEgressStopSearch
+        this.addDefinition(TransitStopByModeAccessEgressStopSearchParameterSet.NAME, TransitStopByModeAccessEgressStopSearchParameterSet::new, () -> this.accessEgressStopSearchParams,
+                params -> this.accessEgressStopSearchParams = (TransitStopByModeAccessEgressStopSearchParameterSet) params);
+
+        this.addDefinition(TransitStopByIdAccessEgressStopSearchParameterSet.NAME, TransitStopByIdAccessEgressStopSearchParameterSet::new, () -> this.accessEgressStopSearchParams,
+                params -> this.accessEgressStopSearchParams = (TransitStopByIdAccessEgressStopSearchParameterSet) params);
+
+        this.addDefinition(CompositeAccessEgressStopSearchParameterSet.NAME, CompositeAccessEgressStopSearchParameterSet::new, () -> this.accessEgressStopSearchParams,
+                params -> this.accessEgressStopSearchParams = (CompositeAccessEgressStopSearchParameterSet) params);
     }
 
     @Parameter
@@ -28,15 +39,21 @@ public class FeederDrtConfigGroup extends ReflectiveConfigGroupWithConfigurableP
     @NotBlank
     public String mode = "feederDrt";
 
+    @Parameter
+    @Comment("A regex that if matches with the from facility id (resp to facility id) will result in an access (resp egress) drt leg not being constructed. Leave empty to consider access and egress segments at all times")
+    public String skipAccessAndEgressAtFacilities;
+
     @NotNull
-    private AccessEgressStopSelectorParams accessEgressStopSelectorConfig;
+    public AccessEgressStopSearchParams accessEgressStopSearchParams;
+
+    public enum AccessEgressStopSelection {CLOSEST}
+
+    @Parameter
+    @NotNull
+    public AccessEgressStopSelection accessEgressStopSelection = AccessEgressStopSelection.CLOSEST;
 
     @Override
     public String getMode() {
         return this.mode;
-    }
-
-    public AccessEgressStopSelectorParams getAccessEgressStopSelectorConfig() {
-        return this.accessEgressStopSelectorConfig;
     }
 }