+++ core/src/main/java/org/eqasim/core/simulation/modes/feeder_drt/utils/AdaptConfigForFeederDrt.java
@@ -10,13 +10,16 @@ import java.util.Set;
 import org.eqasim.core.components.config.EqasimConfigGroup;
 import org.eqasim.core.misc.ClassUtils;
 import org.eqasim.core.simulation.EqasimConfigurator;
+import org.eqasim.core.simulation.mode_choice.constraints.leg_time.LegTimeConstraintConfigGroup;
+import org.eqasim.core.simulation.mode_choice.constraints.leg_time.LegTimeConstraintSingleLegConfigGroup;
 import org.eqasim.core.simulation.modes.drt.utils.AdaptConfigForDrt;
-import org.eqasim.core.simulation.modes.feeder_drt.config.AccessEgressStopSelectorParams;
 import org.eqasim.core.simulation.modes.feeder_drt.config.FeederDrtConfigGroup;
 import org.eqasim.core.simulation.modes.feeder_drt.config.MultiModeFeederDrtConfigGroup;
 import org.eqasim.core.simulation.modes.feeder_drt.mode_choice.EqasimFeederDrtModeChoiceModule;
 import org.eqasim.core.simulation.modes.feeder_drt.mode_choice.constraints.FeederDrtConstraint;
-import org.eqasim.core.simulation.modes.feeder_drt.router.access_egress_selector.ClosestAccessEgressStopSelectorParameterSet;
+import org.eqasim.core.simulation.modes.feeder_drt.router.access_egress_stop_search.CompositeAccessEgressStopSearchParameterSet;
+import org.eqasim.core.simulation.modes.feeder_drt.router.access_egress_stop_search.TransitStopByIdAccessEgressStopSearchParameterSet;
+import org.eqasim.core.simulation.modes.feeder_drt.router.access_egress_stop_search.TransitStopByModeAccessEgressStopSearchParameterSet;
 import org.matsim.contrib.drt.run.MultiModeDrtConfigGroup;
 import org.matsim.contribs.discrete_mode_choice.modules.config.DiscreteModeChoiceConfigGroup;
 import org.matsim.core.config.CommandLine;
@@ -27,7 +30,7 @@ import org.matsim.pt.config.TransitConfigGroup;
 
 public class AdaptConfigForFeederDrt {
 
-    public static void adapt(Config config, Map<String, String> basePtModes, Map<String, String> baseDrtModes, Map<String, String> utilityEstimators, Map<String, String> accessEgressTransitStopModes, String modeAvailability) {
+    public static void adapt(Config config, Map<String, String> basePtModes, Map<String, String> baseDrtModes, Map<String, String> utilityEstimators, Map<String, String> accessEgressTransitStopModes, Map<String, String> accessEgressTransitStopIds, String modeAvailability) {
         if(!config.getModules().containsKey(MultiModeDrtConfigGroup.GROUP_NAME)) {
             throw new IllegalStateException(String.format("Cannot add module '%s' if module '%s' is not present already. You can use '%s' to configure it.", MultiModeFeederDrtConfigGroup.GROUP_NAME, MultiModeDrtConfigGroup.GROUP_NAME, AdaptConfigForDrt.class.getCanonicalName()));
         }
@@ -41,6 +44,11 @@ public class AdaptConfigForFeederDrt {
         TransitConfigGroup transitConfigGroup = (TransitConfigGroup) config.getModules().get(TransitConfigGroup.GROUP_NAME);
         DiscreteModeChoiceConfigGroup dmcConfig = DiscreteModeChoiceConfigGroup.getOrCreate(config);
         EqasimConfigGroup eqasimConfigGroup = EqasimConfigGroup.get(config);
+        LegTimeConstraintConfigGroup legTimeConstraintConfigGroup = (LegTimeConstraintConfigGroup) config.getModules().get(LegTimeConstraintConfigGroup.GROUP_NAME);
+        Map<String, Map<String, LegTimeConstraintSingleLegConfigGroup>> singleLegParameterSetByMainModeByLegMode = null;
+        if(legTimeConstraintConfigGroup != null) {
+            singleLegParameterSetByMainModeByLegMode = legTimeConstraintConfigGroup.getSingleLegParameterSetByMainModeByLegMode();
+        }
 
         // Add DRT to the available modes
         if(modeAvailability != null) {
@@ -66,19 +74,44 @@ public class AdaptConfigForFeederDrt {
                 throw new IllegalStateException(String.format("PT mode '%s' supplied for '%s' is not registered as a transit mode in the '%s' config group", feederDrtConfigGroup.ptModeName, feederDrtMode, TransitConfigGroup.GROUP_NAME));
             }
             feederDrtConfigGroup.accessEgressModeName = baseDrtModes.get(feederDrtMode);
+            for(String mode: new String[]{feederDrtConfigGroup.ptModeName, feederDrtConfigGroup.accessEgressModeName}) {
+                // In this for block, we check if any of the underlying modes are set in the LegTime constraint.
+                // We check for both PT and DRT even though it is highly unlikely for PT, this is so that it already works when we go towards a more generic implementation of intermodality between any modes.
+                if(singleLegParameterSetByMainModeByLegMode != null && singleLegParameterSetByMainModeByLegMode.containsKey(mode) && singleLegParameterSetByMainModeByLegMode.get(mode).containsKey(mode)) {
+                    LegTimeConstraintSingleLegConfigGroup originalLegTimeConstraintSingleLegConfigGroup = singleLegParameterSetByMainModeByLegMode.get(mode).get(mode);
+                    LegTimeConstraintSingleLegConfigGroup legTimeConstraintSingleLegConfigGroup = new LegTimeConstraintSingleLegConfigGroup();
+                    legTimeConstraintSingleLegConfigGroup.mainMode = feederDrtMode;
+                    legTimeConstraintSingleLegConfigGroup.legMode = mode;
+                    legTimeConstraintSingleLegConfigGroup.checkBothDepartureAndArrivalTimes = originalLegTimeConstraintSingleLegConfigGroup.checkBothDepartureAndArrivalTimes;
+                    originalLegTimeConstraintSingleLegConfigGroup.getTimeSlotsParameterSets().forEach(legTimeConstraintSingleLegConfigGroup::addParameterSet);
+                    legTimeConstraintConfigGroup.addParameterSet(legTimeConstraintSingleLegConfigGroup);
+                }
+            }
             if(!drtModes.contains(feederDrtConfigGroup.accessEgressModeName)) {
                 throw new IllegalStateException(String.format("DRT mode '%s' supplied for '%s' is not registered in the '%s' config group. You can use '%s' to configure it", feederDrtConfigGroup.accessEgressModeName, feederDrtMode, MultiModeDrtConfigGroup.GROUP_NAME, AdaptConfigForDrt.class.getCanonicalName()));
             }
 
 
-            ClosestAccessEgressStopSelectorParameterSet closestAccessEgressStopSelectorParameterSet = new ClosestAccessEgressStopSelectorParameterSet();
-            closestAccessEgressStopSelectorParameterSet.accessEgressTransitStopModes = accessEgressTransitStopModes.get(feederDrtMode).replace("|", ",");
-            closestAccessEgressStopSelectorParameterSet.skipAccessAndEgressAtFacilities = "outside";
+            TransitStopByModeAccessEgressStopSearchParameterSet transitStopByModeAccessEgressStopSearchParameterSet = new TransitStopByModeAccessEgressStopSearchParameterSet();
+            transitStopByModeAccessEgressStopSearchParameterSet.accessEgressTransitStopModes = accessEgressTransitStopModes.get(feederDrtMode).replace("|", ",");
 
-            AccessEgressStopSelectorParams accessEgressStopSelectorParams = new AccessEgressStopSelectorParams();
-            accessEgressStopSelectorParams.addParameterSet(closestAccessEgressStopSelectorParameterSet);
+            TransitStopByIdAccessEgressStopSearchParameterSet transitStopByIdAccessEgressStopSearchParameterSet = null;
+            if(accessEgressTransitStopIds.get(feederDrtMode).length() > 0) {
+                transitStopByIdAccessEgressStopSearchParameterSet = new TransitStopByIdAccessEgressStopSearchParameterSet();
+                transitStopByIdAccessEgressStopSearchParameterSet.ids = accessEgressTransitStopIds.get(feederDrtMode).replace("|", ",");
+            }
 
-            feederDrtConfigGroup.addParameterSet(accessEgressStopSelectorParams);
+            if(transitStopByIdAccessEgressStopSearchParameterSet != null) {
+                CompositeAccessEgressStopSearchParameterSet compositeAccessEgressStopSearchParameterSet = new CompositeAccessEgressStopSearchParameterSet();
+                compositeAccessEgressStopSearchParameterSet.addParameterSet(transitStopByModeAccessEgressStopSearchParameterSet);
+                compositeAccessEgressStopSearchParameterSet.addParameterSet(transitStopByIdAccessEgressStopSearchParameterSet);
+                feederDrtConfigGroup.addParameterSet(compositeAccessEgressStopSearchParameterSet);
+            } else {
+                feederDrtConfigGroup.addParameterSet(transitStopByModeAccessEgressStopSearchParameterSet);
+            }
+
+
+            feederDrtConfigGroup.skipAccessAndEgressAtFacilities = "^outside*+";
 
             multiModeFeederDrtConfigGroup.addParameterSet(feederDrtConfigGroup);
 
@@ -99,6 +132,7 @@ public class AdaptConfigForFeederDrt {
                 .allowOptions("base-pt-modes")
                 .allowOptions("estimators")
                 .allowOptions("access-egress-transit-stop-modes")
+                .allowOptions("access-egress-transit-stop-ids")
                 .allowOptions("mode-availability")
                 .allowOptions("configurator-class")
                 .build();
@@ -110,13 +144,14 @@ public class AdaptConfigForFeederDrt {
         String[] basePtModes = cmd.getOption("base-pt-modes").orElse("pt").split(",");
         String[] estimators = cmd.getOption("estimators").orElse(EqasimFeederDrtModeChoiceModule.FEEDER_DRT_ESTIMATOR_NAME).split(",");
         String[] accessEgressTransitStopModes = cmd.getOption("access-egress-transit-stop-modes").orElse("").split(",");
-
+        String[] accessEgressTransitStopIds = cmd.getOption("access-egress-transit-stop-ids").orElse("").split(",");
 
         Map<String, String[]> toExtract = new HashMap<>();
         toExtract.put("base-drt-modes", baseDrtModes);
         toExtract.put("base-pt-modes", basePtModes);
         toExtract.put("estimators", estimators);
         toExtract.put("access-egress-transit-stop-modes", accessEgressTransitStopModes);
+        toExtract.put("access-egress-transit-stop-ids", accessEgressTransitStopIds);
 
         Map<String, Map<String, String>> info = AdaptConfigForDrt.extractDrtInfo(modeNames, toExtract);
 
@@ -130,7 +165,7 @@ public class AdaptConfigForFeederDrt {
         Config config = ConfigUtils.loadConfig(inputConfigPath, configurator.getConfigGroups());
         configurator.addOptionalConfigGroups(config);
 
-        adapt(config, info.get("base-pt-modes"), info.get("base-drt-modes"), info.get("estimators"), info.get("access-egress-transit-stop-modes"), cmd.getOption("mode-availability").orElse(null));
+        adapt(config, info.get("base-pt-modes"), info.get("base-drt-modes"), info.get("estimators"), info.get("access-egress-transit-stop-modes"), info.get("access-egress-transit-stop-ids"), cmd.getOption("mode-availability").orElse(null));
 
         ConfigUtils.writeConfig(config, outputConfigPath);
     }