+++ core/src/main/java/org/eqasim/core/simulation/vdf/io/VDFReaderInterface.java
@@ -1,0 +1,7 @@
+package org.eqasim.core.simulation.vdf.io;
+
+import java.net.URL;
+
+public interface VDFReaderInterface {
+	public void readFile(URL inputFile);
+}