+++ examples/src/main/java/org/eqasim/examples/corsica_vdf/RunCorsicaVDFEngineSimulation.java
@@ -5,16 +5,12 @@ import java.util.HashSet;
 import java.util.Set;
 
 import org.eqasim.core.components.traffic.EqasimTrafficQSimModule;
-import org.eqasim.core.components.transit.EqasimTransitQSimModule;
 import org.eqasim.core.simulation.analysis.EqasimAnalysisModule;
 import org.eqasim.core.simulation.mode_choice.EqasimModeChoiceModule;
 import org.eqasim.ile_de_france.IDFConfigurator;
 import org.eqasim.ile_de_france.mode_choice.IDFModeChoiceModule;
-import org.eqasim.vdf.VDFConfigGroup;
-import org.eqasim.vdf.VDFModule;
-import org.eqasim.vdf.VDFQSimModule;
-import org.eqasim.vdf.engine.VDFEngineConfigGroup;
-import org.eqasim.vdf.engine.VDFEngineModule;
+import org.eqasim.core.simulation.vdf.VDFConfigGroup;
+import org.eqasim.core.simulation.vdf.engine.VDFEngineConfigGroup;
 import org.matsim.api.core.v01.Scenario;
 import org.matsim.core.config.CommandLine;
 import org.matsim.core.config.CommandLine.ConfigurationException;
@@ -77,19 +73,6 @@ public class RunCorsicaVDFEngineSimulation {
 		controller.addOverridingModule(new EqasimAnalysisModule());
 		controller.addOverridingModule(new EqasimModeChoiceModule());
 		controller.addOverridingModule(new IDFModeChoiceModule(cmd));
-
-		// VDF: Add modules
-		controller.addOverridingModule(new VDFModule());
-		controller.addOverridingQSimModule(new VDFQSimModule());
-
-		// VDF Engine: Add modules
-		controller.addOverridingModule(new VDFEngineModule());
-
-		// VDF Engine: Active engine
-		controller.configureQSimComponents(cfg -> {
-			EqasimTransitQSimModule.configure(cfg, controller.getConfig());
-			cfg.addNamedComponent(VDFEngineModule.COMPONENT_NAME); // here
-		});
 
 		controller.run();
 	}