+++ ile_de_france/src/main/java/org/eqasim/ile_de_france/scenario/RunAdaptConfig.java
@@ -8,6 +8,9 @@ import org.matsim.api.core.v01.TransportMode;
 import org.matsim.contribs.discrete_mode_choice.modules.config.DiscreteModeChoiceConfigGroup;
 import org.matsim.core.config.CommandLine.ConfigurationException;
 import org.matsim.core.config.Config;
+import org.matsim.core.config.groups.QSimConfigGroup;
+import org.matsim.core.config.groups.QSimConfigGroup.VehiclesSource;
+import org.matsim.core.config.groups.VehiclesConfigGroup;
 
 public class RunAdaptConfig {
 	static public void main(String[] args) throws ConfigurationException {
@@ -15,7 +18,7 @@ public class RunAdaptConfig {
 		ConfigAdapter.run(args, configurator.getConfigGroups(), RunAdaptConfig::adaptConfiguration);
 	}
 
-	static public void adaptConfiguration(Config config) {
+	static public void adaptConfiguration(Config config, String prefix) {
 		// Adjust eqasim config
 		EqasimConfigGroup eqasimConfig = EqasimConfigGroup.get(config);
 
@@ -37,5 +40,12 @@ public class RunAdaptConfig {
 			config.qsim().setFlowCapFactor(0.045);
 			config.qsim().setStorageCapFactor(0.045);
 		}
+		
+		// Vehicles
+		QSimConfigGroup qsimConfig = config.qsim();
+		qsimConfig.setVehiclesSource(VehiclesSource.fromVehiclesData);
+		
+		VehiclesConfigGroup vehiclesConfig = config.vehicles();
+		vehiclesConfig.setVehiclesFile(prefix + "vehicles.xml.gz");
 	}
 }