+++ vdf/src/main/java/org/eqasim/vdf/io/VDFWriterInterface.java
@@ -1,7 +1,0 @@
-package org.eqasim.vdf.io;
-
-import java.io.File;
-
-public interface VDFWriterInterface {
-	public void writeFile(File outputFile);
-}