+++ CHANGELOG.md
@@ -6,8 +6,11 @@ included in the (note yet determined) next version number.
 
 **Development version**
 
+- The cutters now take a GeoPackage file as an alterative to a ShapeFile
 - Emissions tools have been moved to core package (from ile_de_france)
 - Switched to MATSim 2025 (PR)
+- In switzerland one can now switch off vehicles waiting to enter traffic
+- In swiss module: adjusted the adapt config to allow parametrizing capacity factors and freight in config
 
 **1.5.0**